Decrease test_stdout's expected size: 400 -> 200 KiB
Motivation: Make `cargo test --release` pass. This is useful for NixOS packaging, because NixOS tests against a release binary.

Currently the 'test_stdout' fails with `--release`, since the test expects the PDF to be > 400KiB, but the release PDF is 337 KiB.

Here we see the debug binary outputting 516 KB:

```
$ echo STDOUT | PAPERAGE_PASSPHRASE=secret cargo run  -- -o - | wc --bytes
    Finished dev [unoptimized + debuginfo] target(s) in 0.27s
     Running `target/debug/paper-age -o -`
515997
```

Here we see the release binary outputting 337KB:

```
$ echo STDOUT | PAPERAGE_PASSPHRASE=secret cargo run --release  -- -o - | wc --bytes
    Finished release [optimized] target(s) in 0.41s
     Running `target/release/paper-age -o -`
336980
```

So let's decrease the test limit. I guess the value of the test is mostly in checking that the PDF isn't less than < 10KB.